requirements.txt
PyPDF2>=3.0.1
pdfminer.six==20231228
pdfplumber>=0.10.2
Pillow>=10.0.0
pdf2image>=1.16.3
scholarly>=1.7.11
joblib>=1.4.2